requirements.txt
vosk >= 0.3.45
python-ffmpeg >= 2.0.12
